+++ scripts/end_game.c
@@ -109,7 +109,7 @@ int end_game()
             mouse_pressed(event);
         }
 
-        for (int i = 0; i < sizeof(buttons)/sizeof(buttons[0]); i++)
+        for (int i = 0; i < (int)(sizeof(buttons)/sizeof(buttons[0])); i++)
         {
             out = button_update(buttons[i]);
             switch (out)
@@ -137,7 +137,7 @@ int end_game()
         else
             label_render(&title_label);
         
-        for (int i = 0; i < sizeof(labels)/sizeof(labels[0]); i++)
+        for (int i = 0; i < (int)(sizeof(labels)/sizeof(labels[0])); i++)
         {
             label_update(labels[i]);
             label_render(labels[i]);
