+++ scripts/game.c
@@ -153,7 +153,7 @@ int game()
             }
         }
         
-        for (int i = 0; i < sizeof(labels)/sizeof(labels[0]); i++)
+        for (int i = 0; i < (int)(sizeof(labels)/sizeof(labels[0])); i++)
         {
             label_update(labels[i]);
             label_render(labels[i]);
